Split internal/controller in three files
