+++ internal/controller/controller.go
@@ -17,10 +17,7 @@ limitations under the License.
 package controller
 
 import (
-	"fmt"
-
 	"github.com/golang/glog"
-	"gopkg.in/yaml.v2"
 
 	corev1 "k8s.io/api/core/v1"
 	extensionsv1beta1 "k8s.io/api/extensions/v1beta1"
@@ -110,146 +107,4 @@ func NewLemonLDAPNGController(controllerConfig *Configuration) *LemonLDAPNGContr
 		&corev1.ConfigMap{}, controllerConfig.ResyncPeriod, mapEventHandler)
 
 	return ingressWatcher
-}
-
-// parseIngress returns the ingress namespace, the ingress name, and a map of VHosts
-func (c *LemonLDAPNGController) parseIngress(obj interface{}) (string, string, map[string]*llngconfig.VHost, error) {
-	ingressObj := obj.(*extensionsv1beta1.Ingress)
-	ingressNamespace := ingressObj.Namespace
-	ingressName := ingressObj.Name
-	ingressAnnotations := ingressObj.GetAnnotations()
-	vhosts := make(map[string]*llngconfig.VHost)
-
-	locationRulesAnnotation := "kubernetes-controller.lemonldap-ng.org/location-rules"
-	locationRules := make(map[string]string)
-	locationRulesYaml, ok := ingressAnnotations[locationRulesAnnotation]
-	if ok {
-		err := yaml.Unmarshal([]byte(locationRulesYaml), &locationRules)
-		if err != nil {
-			return ingressNamespace, ingressName, vhosts, fmt.Errorf("Unable to parse locationRules annotation %s of Ingress %s/%s, ignoring Ingress: %s", locationRulesAnnotation, ingressNamespace, ingressName, err)
-		}
-	} else {
-		locationRules = llngconfig.DefaultLocationRules
-	}
-
-	exportedHeadersAnnotation := "kubernetes-controller.lemonldap-ng.org/exported-headers"
-	exportedHeaders := make(map[string]string)
-	exportedHeadersYaml, ok := ingressAnnotations[exportedHeadersAnnotation]
-	if ok {
-		err := yaml.Unmarshal([]byte(exportedHeadersYaml), &exportedHeaders)
-		if err != nil {
-			return ingressNamespace, ingressName, vhosts, fmt.Errorf("Unable to parse exportedHeaders annotation %s of Ingress %s/%s, ignoring Ingress: %s", exportedHeadersAnnotation, ingressNamespace, ingressName, err)
-		}
-	} else {
-		exportedHeaders = llngconfig.DefaultExportedHeaders
-	}
-
-	for _, rule := range ingressObj.Spec.Rules {
-		serverName := rule.Host
-		if serverName == "" || serverName == "*" {
-			serverName = "default"
-		}
-		if rule.HTTP == nil {
-			continue
-		}
-		vhosts[serverName] = llngconfig.NewVHost(serverName, locationRules, exportedHeaders)
-	}
-	return ingressNamespace, ingressName, vhosts, nil
-}
-
-func (c *LemonLDAPNGController) ingressAdded(obj interface{}) {
-	ingressNamespace, ingressName, vhosts, err := c.parseIngress(obj)
-	if err != nil {
-		glog.Error(err)
-		return
-	}
-	glog.Infof("An ingress was created: %s/%s", ingressNamespace, ingressName)
-	c.llngConfig.AddVhosts(vhosts)
-	err = c.llngConfig.Save() // FIXME async + batch
-	if err != nil {
-		glog.Error(err)
-		return
-	}
-}
-
-func (c *LemonLDAPNGController) ingressDeleted(obj interface{}) {
-	ingressNamespace, ingressName, vhosts, err := c.parseIngress(obj)
-	if err != nil {
-		glog.Error(err)
-		return
-	}
-	glog.Infof("An ingress was deleted: %s/%s", ingressNamespace, ingressName)
-	c.llngConfig.DeleteVhosts(vhosts)
-	err = c.llngConfig.Save() // FIXME async + batch
-	if err != nil {
-		glog.Error(err)
-		return
-	}
-}
-
-func (c *LemonLDAPNGController) ingressUpdated(old, cur interface{}) {
-	_, _, oldVhosts, err := c.parseIngress(cur)
-	if err != nil {
-		glog.Error(err)
-		return
-	}
-	curIngressNamespace, curIngressName, curVhosts, err := c.parseIngress(cur)
-	if err != nil {
-		glog.Error(err)
-		return
-	}
-	glog.Infof("An ingress was updated: %s/%s", curIngressNamespace, curIngressName)
-	c.llngConfig.DeleteVhosts(oldVhosts)
-	c.llngConfig.AddVhosts(curVhosts)
-	err = c.llngConfig.Save() // FIXME async + batch
-	if err != nil {
-		glog.Error(err)
-		return
-	}
-}
-
-func (c *LemonLDAPNGController) configMapSmurfed(obj interface{}, verb string) {
-	configMapObj := obj.(*corev1.ConfigMap)
-	configMapKey := fmt.Sprintf("%s/%s", configMapObj.Namespace, configMapObj.Name)
-	if configMapKey == c.controllerConfig.ConfigMapName {
-		glog.Infof("A ConfigMap was %s: %s", verb, configMapKey)
-		if verb == "deleted" {
-			c.llngConfig.SetOverrides(make(map[string]interface{}))
-			err := c.llngConfig.Save() // FIXME async + batch
-			if err != nil {
-				glog.Error(err)
-				return
-			}
-			return
-		}
-		lmConfYaml, ok := configMapObj.Data["lmConf.js"]
-		lmConf := make(map[string]interface{})
-		if !ok {
-			glog.Error("Missing key in ConfigMap: lmConf.js")
-			return
-		}
-		err := yaml.Unmarshal([]byte(lmConfYaml), &lmConf)
-		if err != nil {
-			glog.Errorf("Unable to parse lmConf.js of ConfigMap %s, ignoring Ingress: %s", configMapKey, err)
-			return
-		}
-		c.llngConfig.SetOverrides(lmConf)
-		err = c.llngConfig.Save() // FIXME async + batch
-		if err != nil {
-			glog.Error(err)
-			return
-		}
-	}
-}
-
-func (c *LemonLDAPNGController) configMapAdded(obj interface{}) {
-	c.configMapSmurfed(obj, "added")
-}
-
-func (c *LemonLDAPNGController) configMapDeleted(obj interface{}) {
-	c.configMapSmurfed(obj, "deleted")
-}
-
-func (c *LemonLDAPNGController) configMapUpdated(old, cur interface{}) {
-	c.configMapSmurfed(cur, "updated")
 }
